minor modification in case of setting bits in nil_files in load_csv_fast

+++ OPERATORS/LOAD_CSV/src/load_csv_fast.c
@@ -14,7 +14,6 @@
 //STOP_INCLUDES
 #include "load_csv_fast.h"
 
-
 /*Given a csv file, this method will convert the file into nC binary files
  * where nC is the number of columns in the csv file and each file will
  * contain the contents of a column from the csv file*/
@@ -250,7 +249,7 @@ load_csv_fast(
     //element is not nil, write to not nil buffer
     if ( !is_val_null ) { 
       int8_t bit_idx = row_ctr % 64;
-      nn_buf[col_ctr] |= (1 << bit_idx);
+      nn_buf[col_ctr] |= ((uint64_t)1 << bit_idx);
     }
     else {
       // bit already 0 during initialization so no need to set it to 0
